added christmas-bash
Copy the code from the christmas-bash file and add it to the end of your ~/.bashrc file. Save the file, close it and then open another terminal window or shell. Then, you'll be able to see the changes.  

Want to remove the theme? Comment out / remove the added code. Then, save as normal.  Close and re-open your terminal window for the changes to update.